Renamed the "custom log level" feature to "custom patterns"
This is more reflective of the feature's functionality since it can be used to highlight any word or regex pattern, not just log levels.
Also, since we may decide to allow for other styling the just the color, I adjusted the naming so that the code talks about "decorators",
rather than "colorizers".