New feature - timestamp visualization (#587)
New feature: A time progress indicator is now added to the selected lines when two or more lines are selected. The progress indicator gives a visual representation of each line's point in time within the selection.

The color of the progress indicator can be customized using the logFileHighlighter.progressIndicatorUnderlineColor setting. The default color is #00ff1f8f.

The feature can be disabled using the logFileHighlighter.enableProgressIndicator setting.