Added support for microsecond precision for timestamps  (#603)
Added support for microsecond precision for timestamps in time period duration calculations and timestamp progress indicators.

Example timestamp: 2023-09-26 18:24:17.035169

Before this change, only millisecond precision was supported, so the last three digits in the timestamp above was lost. This was caused by limited time precision support in Javascript and the new microsecond precision has required explicit handling in the extension to keep track of the number of microseconds.

Also did some refactoring and rewriting of unit tests.

Fixes #65 and #594.