Improves progress indicator and error handling
Adds a check to prevent negative decoration character counts in the progress indicator.

Implements try-catch block to improve error handling when decorating lines, displaying an error message to the user.

Adds checks for undefined time and duration properties in TimePeriod and TimeWithMicroseconds classes. This prevents potential errors if these values are not properly initialized.

Fixes #703
Hopefully.